Moved functions and changed dependencies to support the new structure.
