+++ src/main/java/hiof/gruppe1/Estivate/SQLParsers/TextConcatenation/SQLParserTextConcatenation.java
@@ -9,6 +9,7 @@ import hiof.gruppe1.Estivate.objectParsers.IObjectParser;
 import hiof.gruppe1.Estivate.objectParsers.ReflectionParser;
 
 import static hiof.gruppe1.Estivate.SQLAdapters.TableDialectAttributeAdapter.getCompatAttr;
+import static hiof.gruppe1.Estivate.utils.simpleTypeCheck.isSimple;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -23,10 +24,13 @@ public class SQLParserTextConcatenation implements ISQLParser {
     private final WriteBuilder writeBuilder;
     private final ReadBuilder readBuilder;
 
+    private final TextConcatTableManagement tableManagement;
+
     public SQLParserTextConcatenation(IDriverHandler sqlDriver) {
         this.sqlDriver = sqlDriver;
+        this.tableManagement = new TextConcatTableManagement(sqlDriver);
         this.objectParser = new ReflectionParser();
-        this.writeBuilder = new WriteBuilder(sqlDriver, objectParser);
+        this.writeBuilder = new WriteBuilder();
         this.readBuilder = new ReadBuilder();
     }
 
@@ -39,9 +43,52 @@ public class SQLParserTextConcatenation implements ISQLParser {
     }
 
     public Boolean writeToDatabase(SQLWriteObject writeObject) {
-        String writeableString = writeBuilder.createWritableSQLString(writeObject);
+        String writeableString = writeObjectToDatabase(writeObject);
       //  sqlDriver.executeNoReturnSplit(writeableString);
         return true;
+    }
+
+    private String writeObjectToDatabase(SQLWriteObject writeObject) {
+        tableManagement.createOrResizeTableIfNeeded(writeObject);
+
+        if (writeObject.getAttributeList().get("id").getData().toString().equals("0")) {
+            writeObject.getAttributeList().remove("id");
+        }
+
+        String tableName = writeObject.getAttributeList().get("class").getInnerName();
+
+        String insertTable = getObjectClass(writeObject);
+        writeObject.getAttributeList().remove("class");
+
+        // PartBuilders to allow building the String in a non-linear way.
+        // Remove the complex objects after parsing.
+
+        SQLWriteObject writeObjectSimple = new SQLWriteObject();
+        writeObjectSimple.setAttributes((HashMap<String, SQLAttribute>) writeObject.getAttributeList().clone());
+        writeObjectSimple.getAttributeList().entrySet().removeIf(entry -> !isSimple(entry.getValue().getData().getClass()));
+
+        String finalString = writeBuilder.createInsertStatement(tableName, insertTable, writeObjectSimple);
+
+        int parentId = executeGetId(tableName, finalString);
+        traverseNonPrimitives(writeObject, tableName, parentId);
+
+        return String.valueOf(parentId);
+    }
+
+
+    private void traverseNonPrimitives(SQLWriteObject writeObject, String parentNameSimple, int parentId) {
+        writeObject.getAttributeList().forEach((k, v) -> {
+            if (!isSimple(v.getData().getClass())) {
+                HashMap<String, SQLAttribute> parsedAttributes = objectParser.parseObjectToAttributeList(v.getDataRaw());
+                SQLWriteObject recursiveObject = new SQLWriteObject(parsedAttributes);
+                String objectClass = getObjectClass(recursiveObject);
+
+                tableManagement.createAppendingTableIfMissing(parentNameSimple, objectClass, true);
+                String childId = writeObjectToDatabase(recursiveObject);
+                String recursiveRelationship = writeBuilder.createRelationshipInsert(k, parentNameSimple, objectClass, String.valueOf(parentId), childId);
+                sqlDriver.executeNoReturnSplit(recursiveRelationship);
+            }
+        });
     }
 
     public <T> T readFromDatabase(Class<T> castTo, int id) {
@@ -65,7 +112,6 @@ public class SQLParserTextConcatenation implements ISQLParser {
 
         return object;
     }
-
 
     public <T> ArrayList<T> readFromDatabase(Class<T> castTo) {
         String SQLQuery = readBuilder.createReadableSQLString(castTo);
@@ -125,5 +171,17 @@ public class SQLParserTextConcatenation implements ISQLParser {
             return -1;
         }
         return -1;
+    }
+
+    private int executeGetId(String tableName, String executingString) {
+        int id;
+        try {
+            ResultSet rs = sqlDriver.executeQuery(executingString);
+            id = rs.getInt(tableName + "_" + "id");
+            rs.close();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+        return id;
     }
 }
