+++ src/main/java/hiof/gruppe1/Estivate/SQLParsers/TextConcatenation/WriteBuilder.java
@@ -1,74 +1,13 @@
 package hiof.gruppe1.Estivate.SQLParsers.TextConcatenation;
 
-import hiof.gruppe1.Estivate.Objects.SQLAttribute;
 import hiof.gruppe1.Estivate.Objects.SQLWriteObject;
-import hiof.gruppe1.Estivate.drivers.IDriverHandler;
-import hiof.gruppe1.Estivate.objectParsers.IObjectParser;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.HashMap;
-
-import static hiof.gruppe1.Estivate.SQLParsers.TextConcatenation.SQLParserTextConcatenation.getObjectClass;
-import static hiof.gruppe1.Estivate.utils.simpleTypeCheck.isSimple;
 
 public class WriteBuilder {
     private final String INSERT_INTO = "INSERT OR REPLACE INTO ";
     private final String VALUES = " VALUES ";
-    private final TextConcatTableManagement tableManagement;
-    private final IObjectParser objectParser;
-    IDriverHandler driver;
 
-    public WriteBuilder(IDriverHandler driver, IObjectParser objectParser) {
-        this.driver = driver;
-        this.tableManagement = new TextConcatTableManagement(driver);
-        this.objectParser = objectParser;
-    }
-
-    String createWritableSQLString(SQLWriteObject writeObject) {
-        tableManagement.createOrResizeTableIfNeeded(writeObject);
-
-        if (writeObject.getAttributeList().get("id").getData().toString().equals("0")) {
-            writeObject.getAttributeList().remove("id");
-        }
-
-        String tableName = writeObject.getAttributeList().get("class").getInnerName();
-
-        String insertTable = getObjectClass(writeObject);
-        writeObject.getAttributeList().remove("class");
-
-        // PartBuilders to allow building the String in a non-linear way.
-        // Remove the complex objects after parsing.
-
-        SQLWriteObject writeObjectSimple = new SQLWriteObject();
-        writeObjectSimple.setAttributes((HashMap<String, SQLAttribute>) writeObject.getAttributeList().clone());
-        writeObjectSimple.getAttributeList().entrySet().removeIf(entry -> !isSimple(entry.getValue().getData().getClass()));
-
-        String finalString = createInsertStatement(tableName, insertTable, writeObjectSimple);
-
-        int parentId = executeGetId(tableName, finalString);
-        traverseNonPrimitives(writeObject, tableName, parentId);
-
-        return String.valueOf(parentId);
-    }
-
-    private void traverseNonPrimitives(SQLWriteObject writeObject, String parentNameSimple, int parentId) {
-        writeObject.getAttributeList().forEach((k, v) -> {
-            if (!isSimple(v.getData().getClass())) {
-                HashMap<String, SQLAttribute> parsedAttributes = objectParser.parseObjectToAttributeList(v.getDataRaw());
-                SQLWriteObject recursiveObject = new SQLWriteObject(parsedAttributes);
-                String objectClass = getObjectClass(recursiveObject);
-
-                tableManagement.createAppendingTableIfMissing(parentNameSimple, objectClass, true);
-                String childId = createWritableSQLString(recursiveObject);
-                String recursiveRelationship = createRelationshipInsert(k, parentNameSimple, objectClass, String.valueOf(parentId), childId);
-                driver.executeNoReturnSplit(recursiveRelationship);
-            }
-        });
-    }
-
-    private String createInsertStatement(String tableName, String insertTable, SQLWriteObject writeObjectSimple) {
+    String createInsertStatement(String tableName, String insertTable, SQLWriteObject writeObjectSimple) {
         StringBuilder finalString = new StringBuilder();
         StringBuilder keyString = new StringBuilder();
         StringBuilder valuesString = new StringBuilder();
@@ -100,20 +39,8 @@ public class WriteBuilder {
         return finalString.toString();
     }
 
-    private int executeGetId(String tableName, String executingString) {
-        int id;
-        try {
-            ResultSet rs = driver.executeQuery(executingString);
-            id = rs.getInt(tableName + "_" + "id");
-            rs.close();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-        return id;
-    }
 
-
-    private String createRelationshipInsert(String setter, String parentName, String childName, String parentId, String childId) {
+    String createRelationshipInsert(String setter, String parentName, String childName, String parentId, String childId) {
         StringBuilder relationshipInsert = new StringBuilder();
         ArrayList<String> keyValues = new ArrayList<>();
